+++ .gitignore
@@ -1,3 +1,8 @@
+# Developer Specified
+bot_vars
+
+
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
