+++ etc.init.dungeonbot.conf
@@ -11,7 +11,8 @@ script
     mkdir -p /tmp/gunicorn
     cd /home/ubuntu/dungeonbot/
     . env/bin/activate
-    cd dungeonbot/src/
+    . bot_vars
+    cd src/
     gunicorn \
         --bind 0.0.0.0:5006 \
         --log-file /tmp/gunicorn/dungeonbot.log \
