+++ pipreqs.txt
@@ -1,2 +1,8 @@
+click==6.6
+Flask==0.11.1
+itsdangerous==0.24
+Jinja2==2.8
+MarkupSafe==0.23
 requests==2.11.1
 slacker==0.9.24
+Werkzeug==0.11.10
