--- src/helpers.py
@@ -1,6 +1,0 @@
-from __future__ import print_function
-import sys
-
-
-def eprint(*args, **kwargs):
-    print(">>>>", *args, file=sys.stderr, **kwargs)
